requirements.txt
# ============================================
# AI DPR Analysis System - Dependencies
# ============================================

# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0

# --- Database (PostgreSQL) ---
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# --- NLP & Document Processing ---
spacy==3.7.6
pytesseract==0.3.13
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
camelot-py[cv]==0.11.0
tabula-py==2.9.3
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.1.0
langdetect==1.0.9
indic-nlp-library==0.92

# --- Machine Learning ---
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.5.0
shap==0.46.0
lime==0.2.0.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
imbalanced-learn==0.12.3

# --- Data Visualization & Reporting ---
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.0
reportlab==4.2.5

# --- OCR ---
Pillow==10.4.0
opencv-python==4.10.0.84
easyocr==1.7.1

# --- API & Utilities ---
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
websockets==13.0.1
PyJWT==2.10.1

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# --- Logging & Monitoring ---
loguru==0.7.2
prometheus-client==0.21.0
